status dialog as a sepparate component
